JPA 쿼리문 작성

+++ src/main/java/com/acorn/erp/domain/Customer/Repository/AgeGroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.acorn.erp.domain.Customer.Entity.CustomerInfo;
 
-public interface AgeGroupRepository extends JpaRepository<CustomerInfo, Long>{
+public interface AgeGroupRepository extends JpaRepository<CustomerInfo, Integer>{
 
 }
